update nnunet data workflows
