+++ medicai/trainer/nnunet/__init__.py
@@ -1,7 +1,3 @@
-"""
-medicai/trainer/nnunet/__init__.py
-"""
-
 from .pipeline import nnUNetPipeline
 from .training.trainer import nnUNetTrainer
